Move code in cogs
